pyproject.toml
[project]
name = "aoacloudpunchio"
version = "0.1.0"
description = "Automated punch clock system for AoaCloud ERP"
requires-python = ">=3.11"
dependencies = [
    "playwright>=1.45.0",
    "apscheduler>=3.10.4",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "loguru>=0.7.0",
    "aiohttp>=3.8.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "mypy>=1.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.black]
line-length = 88
target-version = ['py311']

[tool.isort]
profile = "black"

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true

[dependency-groups]
dev = [
    "mypy>=1.17.1",
    "ruff>=0.12.9",
]
